rtt_ros: Moving global functions into header, ditching service requester for global service
